Try to clean up the logging view code
- the `LogViewAppender` does way too much UI work and can be replaced
with a simple log appender implementation (inlined)
- the UI-related code was moved to a new `LogView` helper class
providing a static method for setting up the table (not sure about this
part)
- the helper is called exactly once during init of
`ApplicationController`
